requirements.txt
streamlit 
groq 
icalendar 
python-dotenv
tzdata